# storage/docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Week 13 Assignment API",
        "version": "1.0.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/posts": {
            "get": {
                "tags": [
                    "Post"
                ],
                "summary": "Get post list",
                "description": "Get post list",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Provide in header request: Authorization: Bearer ACCESS_TOKEN",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "title",
                        "in": "query",
                        "description": "Filter by post title",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Post list",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "properties": {
                                        "post": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/Post"
                                            }
                                        }
                                    },
                                    "type": "object"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            },
            "post": {
                "tags": [
                    "Post"
                ],
                "summary": "Create new post",
                "description": "Create new post",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Provide in header request: Authorization: Bearer ACCESS_TOKEN",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "title",
                        "in": "formData",
                        "description": "title",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "category_id",
                        "in": "formData",
                        "description": "category_id",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "content",
                        "in": "formData",
                        "description": "content",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Post was stored successfully.",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "post": {
                                    "type": "object",
                                    "$ref": "#/definitions/Post"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Missing required field"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
        "/posts/{id}": {
            "put": {
                "tags": [
                    "Post"
                ],
                "summary": "Update post",
                "description": "Update post",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Provide in header request: Authorization: Bearer ACCESS_TOKEN",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "post id",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "title",
                        "in": "formData",
                        "description": "title",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "category_id",
                        "in": "formData",
                        "description": "category_id",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "content",
                        "in": "formData",
                        "description": "content",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Post was updated successfully.",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "type": "post",
                                    "$ref": "#/definitions/Post"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Missing required field"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Post"
                ],
                "summary": "Delete post",
                "description": "Delete post",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Provide in header request: Authorization: Bearer ACCESS_TOKEN",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "post id",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Post was deleted successfully.",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "type": "object"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        }
    },
    "definitions": {
        "Category": {
            "required": [
                "label",
                "name"
            ],
            "properties": {
                "id": {
                    "description": "id",
                    "type": "integer",
                    "format": "int32"
                },
                "name": {
                    "description": "name",
                    "type": "string"
                },
                "created_at": {
                    "description": "created_at",
                    "type": "string",
                    "format": "date-time"
                },
                "updated_at": {
                    "description": "updated_at",
                    "type": "string",
                    "format": "date-time"
                }
            }
        },
        "Post": {
            "required": [
                "title",
                "category_id",
                "creator_id"
            ],
            "properties": {
                "id": {
                    "description": "id",
                    "type": "integer",
                    "format": "int32"
                },
                "title": {
                    "description": "title",
                    "type": "string"
                },
                "category": {
                    "description": "category",
                    "type": "category",
                    "$ref": "#/definitions/Category"
                },
                "author": {
                    "description": "author",
                    "type": "string"
                },
                "content": {
                    "description": "content",
                    "type": "string"
                },
                "creator_id": {
                    "description": "creator_id",
                    "type": "string"
                },
                "created_at": {
                    "description": "created_at",
                    "type": "string",
                    "format": "date-time"
                },
                "updated_at": {
                    "description": "updated_at",
                    "type": "string",
                    "format": "date-time"
                }
            }
        }
    }
}